fixed dropdown menu, added delete button

+++ back/UserService/UserService/UserEndpoints.cs
@@ -6,6 +6,7 @@ using RabbitMQ.Client.Events;
 using UserService.Models;
 using UserService.Services;
 using System.IdentityModel.Tokens.Jwt;
+using Microsoft.AspNetCore.Http;
 
 
 namespace UserService;
@@ -73,13 +74,42 @@ public static class UserEndpoints
             }
 
             var userId = userIdClaim.Value;
-
+            
             var favorites = await userService.GetUsersWatchStatusAsync(movieId, userId);
+            if (favorites == null)
+            {
+                return Results.Ok("");
+            }
             return Results.Ok(favorites);
         }).WithOpenApi();
 
-        app.MapPost("/delete", async (IUserService userService, UserMovieDeleteDTO delete) =>
+        app.MapPost("/favorites/delete", async (HttpContext httpContext, IUserService userService, UserMovieDeleteDTO delete) =>
         {
+            var authHeader = httpContext.Request.Headers.Authorization.FirstOrDefault();
+            if (string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+            {
+                return Results.Unauthorized();
+            }
+
+            var token = authHeader.Substring("Bearer ".Length).Trim();
+
+            var handler = new JwtSecurityTokenHandler();
+            if (!handler.CanReadToken(token))
+            {
+                return Results.Unauthorized();
+            }
+
+            var jwtToken = handler.ReadJwtToken(token);
+            var userIdClaim = jwtToken.Claims.FirstOrDefault(c =>
+            c.Type == JwtRegisteredClaimNames.Sub || c.Type == "nameid");
+
+            if (userIdClaim == null)
+            {
+                return Results.Unauthorized();
+            }
+
+            var userId = userIdClaim.Value;
+
             var deleted = await userService.DeleteFavoriteMovieAsync(delete);
             return Results.Ok(deleted);
         }).WithOpenApi();
